pyproject.toml
[project]
name = "soji-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=1.0.1",
    "loguru>=0.7.3",
    "openai>=2.21.0",
    "paddleocr>=3.4.0",
    "paddlepaddle-gpu==3.3.0",
    "paddlex[ocr]>=3.4.2",
    "pdf2image>=1.17.0",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "tabulate>=0.9.0",
]

[[tool.uv.index]]
name = "paddle-cu130"
url = "https://www.paddlepaddle.org.cn/packages/stable/cu130/"
explicit = true

[tool.uv.sources]
paddlepaddle-gpu = { index = "paddle-cu130" }
